Embrace Pages, Restructure Repo, Performance
Pages are a thing in next so lets utilize them and while we are at it, why not use the dynamic route api to populate more data!

Took the liberty of moving some files around that would better support build tooling such as rollup for  a component/design library and css preprocessor like less or sass.

Implemented a couple quick performant touches to load things not in view as the user is interacting with the page and added navigable links back to the get started/registration